requirements.txt
# === Núcleo do projeto ===
pandas>=2.0.0
numpy>=1.24.0

# === Utilitários e data wrangling ===
openpyxl>=3.1.0        # leitura/escrita de Excel
pyarrow>=15.0.0        # compatibilidade com parquet/csv.gz
python-dateutil>=2.8.2
tqdm>=4.66.0           # barras de progresso (opcional)

# === Visualização e dashboards (opcional, mas recomendável) ===
matplotlib>=3.8.0
plotly>=5.20.0
streamlit>=1.36.0

# === Machine Learning / Clustering (para fases futuras) ===
scikit-learn>=1.4.0
scipy>=1.12.0

# === Geodados (para integração futura com mapas, se for usar) ===
geopandas>=0.14.0
folium>=0.15.0

# === Diversos ===
unicodedata2>=15.1.0   # normalização de acentos (se não estiver no core)

Pillow>=10.0.0

pyarrow==16.1.0

